Move Clean in the spirit of #3192
Backward-compatibility shim is in place.
